change interface back to Button component
