+++ src/components/ui/Button.tsx
@@ -1,5 +1,14 @@
-import { ButtonColor, ButtonProps } from "@/interface/button";
 import CloseIcon from "./CloseIcon";
+
+type ButtonColor = "primary" | "secondary" | "ternary" | "warning" | "info" | "light" | "dark";
+export type Props = {
+  color?: ButtonColor;
+  size?: "tiny" | "small" | "normal" | "large";
+  outline?: boolean;
+  variant?: "normal" | "icon";
+  label?: string;
+  onClick?: () => void;
+};
 
 const btn = " text-light text-base rounded-full font-[Roboto] flex justify-center items-center";
 const getButtonClass = (color: ButtonColor, outline: boolean) => {
@@ -53,7 +62,7 @@ const Button = ({
   onClick,
   label = "button",
   ...otherProps
-}: ButtonProps) => {
+}: React.ButtonHTMLAttributes<HTMLButtonElement> & Props) => {
   const buttonColor = {
     primary: getButtonClass("primary", outline),
     secondary: getButtonClass("secondary", outline),
